컨트롤러 수정_몇번째인지 모르겟음3

+++ src/main/java/univ/lecture/riotapi/model/Summoner.java
@@ -14,7 +14,7 @@ public class Summoner {
 	/* private String name;
     private int summonerLevel;*/
 	private int teamId;
-	private int now;
+	private long now;
 	private double result;
